[no ci] [WIP] add custom glowing color split health into multiple files
