+++ src/main/resources/usbplus.mixins.json
@@ -3,7 +3,8 @@
   "package": "net.usbwire.usbplus.mixins",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-		"ChatHudMixin"
+		"ChatHudMixin",
+		"EntityMixin"
 	],
   "injectors": {
     "defaultRequire": 1
